Fix level screen
